Extract presentational components into @sphere/ui workspace package
