+++ apps/sphere/src/components/App/index.tsx
@@ -4,10 +4,10 @@ import { actions, selectors } from "@/store"
 import { selectShowLeftSidebar } from "@/store/app"
 import { useAppDispatch, useAppSelector } from "@/store/hooks"
 import addFromClipboard from "@/store/source/addFromClipboard"
-import { AppLayout } from "@/ui/AppLayout"
-import { ErrorFallback } from "@/ui/ErrorFallback"
-import { Overlay } from "@/ui/Overlay"
-import { Sidebar } from "@/ui/Sidebar"
+import { AppLayout } from "@sphere/ui/AppLayout"
+import { ErrorFallback } from "@sphere/ui/ErrorFallback"
+import { Overlay } from "@sphere/ui/Overlay"
+import { Sidebar } from "@sphere/ui/Sidebar"
 import { Center } from "@mantine/core"
 import { useHotkeys } from "@mantine/hooks"
 import { useCallback } from "react"
