+++ apps/sphere/src/components/LayerPanel/index.tsx
@@ -2,7 +2,7 @@ import { isRasterTileFormat } from "@/lib/tilejson"
 import { actions, selectors } from "@/store"
 import { useAppDispatch } from "@/store/hooks"
 import { LayerType, SourceType } from "@/types"
-import { ActionBar } from "@/ui/ActionBar"
+import { ActionBar } from "@sphere/ui/ActionBar"
 import { Flex, Select, TextInput } from "@mantine/core"
 import { createSelector } from "@reduxjs/toolkit"
 import { IconCopy, IconCrosshair, IconTrash } from "@tabler/icons"
