+++ apps/sphere/src/components/LeftSidebar/SourcesTab.tsx
@@ -2,7 +2,7 @@ import { actions, selectors } from "@/store"
 import { useAppDispatch, useAppSelector } from "@/store/hooks"
 import logger from "@/logger"
 import { SourceType } from "@/types"
-import { ActionBar, type ActionBarOnClick } from "@/ui/ActionBar"
+import { ActionBar, type ActionBarOnClick } from "@sphere/ui/ActionBar"
 import { Accordion, Button, Group, Modal, TextInput } from "@mantine/core"
 import { useForm } from "@mantine/form"
 import { IconCrosshair, IconFile, IconLink, IconPlus, IconTrash } from "@tabler/icons"
