+++ apps/sphere/src/components/MapContextMenu/index.tsx
@@ -2,7 +2,7 @@ import { useCursor } from "@/hooks/useCursor"
 import { copySelectionAsGeojson, copySelectionAsWkt } from "@/lib/copy-selection"
 import { useAppSelector } from "@/store/hooks"
 import { selectors } from "@/store/selectors"
-import { ContextMenu } from "@/ui/ContextMenu"
+import { ContextMenu } from "@sphere/ui/ContextMenu"
 import { CopyButton, Menu, Text } from "@mantine/core"
 import { IconCopy, IconSearch } from "@tabler/icons"
 import { useCallback, useEffect, useState } from "react"
