+++ apps/sphere/src/components/SourcesOutline/index.tsx
@@ -1,6 +1,6 @@
 import { actions, selectors } from "@/store"
-import { Outline, type OutlineOnMove, type OutlineRenderItem } from "@/ui/Outline"
-import { OutlineItem } from "@/ui/Outline/OutlineItem"
+import { Outline, type OutlineOnMove, type OutlineRenderItem } from "@sphere/ui/Outline"
+import { OutlineItem } from "@sphere/ui/Outline/OutlineItem"
 import { useMantineTheme } from "@mantine/core"
 import { createSelector } from "@reduxjs/toolkit"
 import { IconBraces } from "@tabler/icons"
