+++ apps/sphere/src/components/SphereMap/Draw.tsx
@@ -5,7 +5,7 @@ import { loadDrawFeatures } from "@/lib/draw-source"
 import logger from "@/logger"
 import { actions } from "@/store"
 import { useAppDispatch, useAppSelector } from "@/store/hooks"
-import { Overlay } from "@/ui/Overlay"
+import { Overlay } from "@sphere/ui/Overlay"
 import { Button, Flex } from "@mantine/core"
 import { useCallback, useEffect, useRef, useState } from "react"
 import { useMap } from "react-map-gl/maplibre"
