+++ apps/sphere/src/main.tsx
@@ -7,7 +7,7 @@ import { setupMaplibre } from "@/maplibre"
 import { store } from "@/store"
 import { handleDragDrop, handleHotkey, handleTheme, handleVersion } from "@/tauri"
 import { setupKeyboard } from "@/store/keyboard"
-import { RootErrorFallback } from "@/ui/ErrorFallback/RootErrorFallback"
+import { RootErrorFallback } from "@sphere/ui/ErrorFallback/RootErrorFallback"
 import React from "react"
 import ReactDOM from "react-dom/client"
 import { ErrorBoundary } from "react-error-boundary"
