+++ packages/ui/biome.json
@@ -1,0 +1,27 @@
+{
+    "$schema": "https://biomejs.dev/schemas/1.9.0/schema.json",
+    "extends": ["../../biome.json"],
+    "linter": {
+        "rules": {
+            "nursery": {
+                "noRestrictedImports": {
+                    "level": "error",
+                    "options": {
+                        "paths": {
+                            "@/*": "Packages must not import from the app alias (@/\u2026).",
+                            "@tauri-apps/api": "@sphere/ui must not depend on Tauri.",
+                            "@reduxjs/toolkit": "@sphere/ui must not depend on Redux.",
+                            "react-redux": "@sphere/ui must not depend on Redux.",
+                            "maplibre-gl": "@sphere/ui must not depend on maplibre-gl.",
+                            "@maplibre/maplibre-gl-style-spec": "@sphere/ui must not depend on maplibre.",
+                            "react-map-gl": "@sphere/ui must not depend on react-map-gl."
+                        }
+                    }
+                }
+            }
+        }
+    },
+    "files": {
+        "include": ["src/**/*.ts", "src/**/*.tsx"]
+    }
+}
