+++ packages/ui/src/PropertiesTable/BarChart.tsx
