+++ packages/ui/src/PropertiesViewer/CopyButton.tsx
