+++ packages/ui/tsconfig.json
@@ -1,0 +1,12 @@
+{
+    "extends": "../../tsconfig.base.json",
+    "compilerOptions": {
+        "baseUrl": ".",
+
+        "paths": {
+            "@sphere/utils/*": ["../utils/src/*"],
+            "@tabler/icons": ["../../node_modules/@tabler/icons/icons-react"]
+        }
+    },
+    "include": ["src"]
+}
