+++ packages/ui/vitest.config.ts
@@ -1,0 +1,13 @@
+/// <reference types="vitest" />
+import { defineConfig } from "vitest/config"
+import react from "@vitejs/plugin-react"
+
+export default defineConfig({
+    plugins: [react()],
+    test: {
+        include: ["src/**/*.test.ts", "src/**/*.test.tsx"],
+        globals: true,
+        environment: "happy-dom",
+        setupFiles: ["./src/setupTests.ts"],
+    },
+})
